dashboard/requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
